requirements.txt
setuptools==59.5.0
lvis
scipy

# Detectron2 dependencies; note: we installed pytorch/cudatoolkit using conda, here we just list all the packages for reference
pytorch==1.10.1
torchvision==0.11.2
torchaudio==0.10.1
cudatoolkit==11.3
--find-links https://dl.fbaipublicfiles.com/detectron2/wheels/cu113/torch1.10/index.html
detectron2